pyproject.toml
[project]
name = "roost-rental"
version = "0.1.0"
description = "Roost — Self-hosted vacation rental operations platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "sqlalchemy>=2.0",
    "alembic",
    "pydantic-settings[yaml]",
    "jinja2",
    "typer",
    "questionary",
    "structlog",
    "httpx",
    "psycopg[binary]",
    "python-slugify",
    "polars",
    "python-multipart",
    "pypdf>=4.0",
    "aiosmtplib>=5.1.0",
    "apscheduler>=3.11.2",
    "tenacity>=9.1.4",
    "ollama",
    "sse-starlette",
    "sqlglot",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-dateutil",
]

[project.scripts]
manage = "manage:cli"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
